bump the constant limit

+++ src/compiler.c
@@ -256,14 +256,14 @@ static uint8_t makeConstant(Value value) {
   }
 
   int constant = addConstant(currentChunk(), value);
-  if (constant > UINT8_MAX) {
+  if (constant > UINT16_MAX) {
     error("Too many constants in one chunk.");
     return 0;
   }
 
   if (hashable) mapSet(current->constants, value, NUMBER_VAL(constant));
 
-  return (uint8_t)constant;
+  return (uint16_t)constant;
 }
 
 static void loadConstant(Value value) {
